requirements/dev.txt
-r prod.txt

# Testing
nose==1.3.4
Flask-Testing==0.4.2
coverage==3.7.1
mock==1.0.1
